utils: Silence an MSVC warning about conversion from double to float
This silences the following MSVC warning:

    ../src/utils.c(318): warning C4244: 'initializing': conversion from 'double' to 'const float', possible loss of data

A simpler fix would be to just change the 1.0 constant into a float
constant 1.0f. This however causes warnings with Clang:

    ../src/utils.c:318:34: warning: implicit conversion from 'unsigned int' to 'float' changes value from 4294967295 to 4294967296 [-Wimplicit-const-int-float-conversion]
        const float scale = range / (UINT32_MAX + 1.0f);
                                     ^~~~~~~~~~ ~

+++ src/utils.c
@@ -315,7 +315,7 @@ void checkasm_randomize_range(double *buf, int width, double range)
 
 void checkasm_randomize_rangef(float *buf, int width, float range)
 {
-    const float scale = range / (UINT32_MAX + 1.0);
+    const float scale = (float) (range / (UINT32_MAX + 1.0));
     while (width--)
         *buf++ = scale * checkasm_rand_uint32();
 }
